\name{mosaic.importToolboxSpssDataFile}
\alias{mosaic.importToolboxSpssDataFile}

\title{
importToolboxSpssDataFile
}
\description{
load dat-file from 'toolbox for resarch' spss export with tab-separator with n columns to dataframe
}
\usage{
mosaic.importToolboxSpssDataFile(filename)
}
\arguments{
  \item{filename}{
filename or a complete path to a dat-file
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}

\keyword{ load }
\keyword{ spss }
\keyword{ toolbox for research }
\keyword{ data }